+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.7.0 - TBD
+
+### Added
+
+- [#19](https://github.com/zendframework/zend-paginator/pull/19) adds:
+  - `Zend\Paginator\AdapterPluginManagerFactory`
+  - `Zend\Paginator\ScrollingStylePluginManagerFactory`
+  - `ConfigProvider`, which maps the `AdapterPluginManager` and
+    `ScrollingStylePluginManager` services to the above factories.
+  - `Module`, which does the same, for zend-mvc contexts.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
 ## 2.6.1 - 2016-04-11
 
 ### Added